feat(java/driver/jni): support loading JNI library from external path
Allow embedders to load the JNI native library from a custom path by
setting the `arrow.adbc.driver.jni.library.path` JVM system property.
When set, JniLoader tries to load the library from that directory first,
falling back to JAR extraction if the file is not found. This is
consistent with the approach used by arrow-java's JniLoader.

Add JniLibraryResolver to separate path resolution from native loading,
enabling unit testing without triggering JNI class initialization.

Closes apache/arrow-adbc#4207
